Combine ConsoleWriter and ConsoleReader into Console
@maikon
@jsuchy
